requirements.txt
flask
transformers
sentencepiece
accelerate
datasets
evaluate
rouge_score
torchvision
torchaudio
torch
python-dotenv